Avoid conversion we don't need
cffi can cope with these being doubles.

+++ sb_vision/native/cv3d_build.py
@@ -19,8 +19,8 @@ CV3D_DECLS = """
         const float image_points[],
         const float camera_matrix[],
         const float dist_coeffs[],
-        float rvec[],
-        float tvec[]
+        double rvec[],
+        double tvec[]
     );
 """
